requirements.txt
six
flask
tqdm
requests
paramiko
pywebio
pyserial
flask-cors
python-vlc
tkinterweb
pywebview
openai
streamlit
scp
scpclient
pydub 
ffmpeg
Pillow
image
xmltodict
netifaces
tkinterdnd2
pygame